# Add `ElementParser` and probably slightly increase performance