--- a/src/reader/buffered_reader.rs
+++ b/src/reader/buffered_reader.rs
@@ -101,31 +101,31 @@ macro_rules! impl_buffered_source {
             let mut read = 0;
             let start = buf.len();
             loop {
-                let used = {
-                    let available = match self $(.$reader)? .fill_buf() $(.$await)? {
-                        Ok(n) if n.is_empty() => break,
-                        Ok(n) => n,
-                        Err(ref e) if e.kind() == io::ErrorKind::Interrupted => continue,
-                        Err(e) => {
-                            *position += read;
-                            return Err(Error::Io(e.into()));
-                        }
-                    };
-
-                    if let Some(i) = parser.feed(available) {
-                        buf.extend_from_slice(&available[..i]);
-
-                        // +1 for `>` which we do not include
-                        self $(.$reader)? .consume(i + 1);
-                        read += i + 1;
-
+                let available = match self $(.$reader)? .fill_buf() $(.$await)? {
+                    Ok(n) if n.is_empty() => break,
+                    Ok(n) => n,
+                    Err(ref e) if e.kind() == io::ErrorKind::Interrupted => continue,
+                    Err(e) => {
                         *position += read;
-                        return Ok(&buf[start..]);
+                        return Err(Error::Io(e.into()));
                     }
-
-                    buf.extend_from_slice(available);
-                    available.len()
                 };
+
+                if let Some(i) = parser.feed(available) {
+                    buf.extend_from_slice(&available[..i]);
+
+                    // +1 for `>` which we do not include
+                    self $(.$reader)? .consume(i + 1);
+                    read += i + 1;
+
+                    *position += read;
+                    return Ok(&buf[start..]);
+                }
+
+                // The `>` symbol not yet found, continue reading
+                buf.extend_from_slice(available);
+
+                let used = available.len();
                 self $(.$reader)? .consume(used);
                 read += used;
             }